feat: move setMetadata logic to getUploadInfo
I haven't been able to reproduce so far, but I've noticed there are objects
in s3 that don't exist in the db. Every invocation of setMetadata is logged,
so I confirmed there have been multiple occurrences where
getRequiredUploadInfo was not followed by a call to setMetadata. This
was brought to my attention when the China team reported they
received an empty result for parameters that should have a match.
The only explanation I can think of is that the page is being closed
after the upload but before setMetadata is called.
Unfortunately, we're unable to view client logs in AWS, so there's
no records of failed upload attempts.